refactor: extract toast to its own component
